Refactor company interest routes to `react-router` v6
Includes rewriting components such as `CompanyInterestList` to a
functional component, and semesters GUI form to `react-final-form`.
