+++ app/routes/companyInterest/components/CompanyInterestList.tsx
@@ -1,18 +1,30 @@
 import { Button, Flex, Icon } from '@webkom/lego-bricks';
-import { Component } from 'react';
-import { Link } from 'react-router-dom';
+import qs from 'qs';
+import { useEffect, useMemo, useState } from 'react';
+import { useLocation, Link, useNavigate } from 'react-router-dom-v5-compat';
 import Select from 'react-select';
+import { fetchSemesters } from 'app/actions/CompanyActions';
+import {
+  deleteCompanyInterest,
+  fetch,
+  fetchAll,
+} from 'app/actions/CompanyInterestActions';
 import { Content } from 'app/components/Content';
 import { selectTheme, selectStyles } from 'app/components/Form/SelectInput';
 import { ConfirmModal } from 'app/components/Modal/ConfirmModal';
 import Table from 'app/components/Table';
 import Tooltip from 'app/components/Tooltip';
-import type { CompanyInterestEntity } from 'app/reducers/companyInterest';
-import type { CompanySemesterEntity } from 'app/reducers/companySemesters';
+import { selectCompanyInterestList } from 'app/reducers/companyInterest';
+import {
+  selectCompanySemestersForInterestForm,
+  type CompanySemesterEntity,
+} from 'app/reducers/companySemesters';
 import { ListNavigation } from 'app/routes/bdb/utils';
-import { getCsvUrl, semesterToText } from '../utils';
+import { useAppDispatch, useAppSelector } from 'app/store/hooks';
+import { EVENT_TYPE_OPTIONS, getCsvUrl, semesterToText } from '../utils';
 import styles from './CompanyInterest.css';
-import { EVENT_TYPE_OPTIONS } from './CompanyInterestPage';
+import { selectIsLoggedIn } from 'app/reducers/auth';
+import { LoginPage } from 'app/components/LoginForm';
 
 type SemesterOptionType = {
   id: number;
@@ -26,254 +38,284 @@ type EventOptionType = {
   label: string;
 };
 
-type Props = {
-  companyInterestList: Array<CompanyInterestEntity>;
-  deleteCompanyInterest: (arg0: number) => Promise<any>;
-  fetch: (arg0: Record<string, any>) => Promise<any>;
-  hasMore: boolean;
-  fetching: boolean;
-  semesters: Array<CompanySemesterEntity>;
-  replace: (arg0: string) => void;
-  selectedSemesterOption: SemesterOptionType;
-  selectedEventOption: EventOptionType;
-  router: any;
-  authToken: string;
-};
+const CompanyInterestList = () => {
+  const [generatedCSV, setGeneratedCSV] = useState<
+    { url: string; filename: string } | undefined
+  >(undefined);
 
-type State = {
-  generatedCSV?: { url: string; filename: string };
-};
-
-class CompanyInterestList extends Component<Props, State> {
-  state = {
-    generatedCSV: undefined,
+  const location = useLocation();
+  const semesterId = Number(
+    qs.parse(location.search, {
+      ignoreQueryPrefix: true,
+    }).semesters
+  );
+  const semesters = useAppSelector((state) =>
+    selectCompanySemestersForInterestForm(state)
+  );
+  const semesterObj: CompanySemesterEntity | null | undefined = semesters.find(
+    (semester) => semester.id === semesterId
+  );
+  const eventValue = qs.parse(location.search, {
+    ignoreQueryPrefix: true,
+  }).event;
+  const selectedSemesterOption = useMemo(
+    () => ({
+      id: semesterId ? semesterId : 0,
+      semester: semesterObj != null ? semesterObj.semester : '',
+      year: semesterObj != null ? semesterObj.year : '',
+      label:
+        semesterObj != null
+          ? semesterToText({
+              semester: semesterObj.semester,
+              year: semesterObj.year,
+              language: 'norwegian',
+            })
+          : 'Vis alle semestre',
+    }),
+    [semesterId, semesterObj]
+  );
+  const selectedEventOption = {
+    value: eventValue ? eventValue : '',
+    label: eventValue
+      ? EVENT_TYPE_OPTIONS.find((eventType) => eventType.value === eventValue)
+          .label
+      : 'Vis alle arrangementstyper',
   };
+  const companyInterestList = useAppSelector((state) =>
+    selectCompanyInterestList(
+      state,
+      selectedSemesterOption.id,
+      selectedEventOption.value
+    )
+  );
+  const hasMore = useAppSelector((state) => state.companyInterest.hasMore);
+  const fetching = useAppSelector((state) => state.companyInterest.fetching);
+  const authToken = useAppSelector((state) => state.auth.token);
 
-  handleSemesterChange = (clickedOption: SemesterOptionType): void => {
+  const dispatch = useAppDispatch();
+  const navigate = useNavigate();
+
+  const handleSemesterChange = (clickedOption: SemesterOptionType): void => {
     const { id } = clickedOption;
-    this.props
-      .fetch({
+    dispatch(
+      fetch({
         filters: {
           semesters: id !== null ? id : null,
         },
       })
-      .then(() => {
-        this.props.replace(
-          `/companyInterest?semesters=${clickedOption.id}&event=${this.props.selectedEventOption.value}`
-        );
-      });
+    ).then(() => {
+      navigate(
+        `/companyInterest?semesters=${clickedOption.id}&event=${selectedEventOption.value}`,
+        { replace: true }
+      );
+    });
   };
 
-  handleEventChange = (clickedOption: EventOptionType): void => {
-    this.props.replace(
-      `/companyInterest?semesters=${this.props.selectedSemesterOption.id}&event=${clickedOption.value}`
+  const handleEventChange = (clickedOption: EventOptionType): void => {
+    navigate(
+      `/companyInterest?semesters=${selectedSemesterOption.id}&event=${clickedOption.value}`,
+      { replace: true }
     );
   };
 
-  componentDidUpdate(prevProps) {
-    if (
-      this.props.selectedSemesterOption !== prevProps.selectedSemesterOption
-    ) {
-      this.setState({
-        generatedCSV: undefined,
-      });
-    }
+  useEffect(() => {
+    setGeneratedCSV(undefined);
+  }, [selectedSemesterOption]);
+
+  useEffect(() => {
+    dispatch(fetchAll());
+    dispatch(fetchSemesters());
+  }, [dispatch]);
+
+  const loggedIn = useAppSelector((state) => selectIsLoggedIn(state));
+  if (!loggedIn) {
+    return LoginPage;
   }
 
-  async exportInterestList(event?: string) {
-    const blob = await fetch(
-      getCsvUrl(
-        this.props.selectedSemesterOption.year,
-        this.props.selectedSemesterOption.semester,
-        event
-      ),
-      {
-        headers: {
-          Authorization: `Bearer ${this.props.authToken}`,
-        },
-      }
+  const exportInterestList = async (event?: string) => {
+    const blob = await dispatch(
+      fetch(
+        getCsvUrl(
+          selectedSemesterOption.year,
+          selectedSemesterOption.semester,
+          event
+        ),
+        {
+          headers: {
+            Authorization: `Bearer ${authToken}`,
+          },
+        }
+      )
     ).then((response) => response.blob());
     return {
       url: URL.createObjectURL(blob),
-      filename: `company-interests-${this.props.selectedSemesterOption.year}-${
-        this.props.selectedSemesterOption.semester
-      }${
-        this.props.selectedEventOption.value
-          ? `-${this.props.selectedEventOption.value}`
-          : ''
-      }.csv`,
+      filename: `company-interests-${selectedSemesterOption.year}-${
+        selectedSemesterOption.semester
+      }${selectedEventOption.value ? `-${selectedEventOption.value}` : ''}.csv`,
     };
-  }
+  };
 
-  render() {
-    const { generatedCSV } = this.state;
-    const columns = [
-      {
-        title: 'Bedriftsnavn',
-        dataIndex: 'companyName',
-        render: (companyName: string, companyInterest: Record<string, any>) => (
-          <Link to={`/companyInterest/${companyInterest.id}/edit`}>
-            {companyInterest.company
-              ? companyInterest.company.name
-              : companyName}
-          </Link>
-        ),
-      },
-      {
-        title: 'Kontaktperson',
-        dataIndex: 'contactPerson',
-        render: (contactPerson: string) => <span>{contactPerson}</span>,
-      },
-      {
-        title: 'E-post',
-        dataIndex: 'mail',
-        render: (mail: string) => <span>{mail}</span>,
-      },
-      {
-        title: '',
-        dataIndex: 'id',
-        render: (id) => (
-          <Flex justifyContent="center">
-            <ConfirmModal
-              title="Bekreft fjerning av interesse"
-              message={'Er du sikker på at du vil fjerne?'}
-              closeOnConfirm
-              onConfirm={() => this.props.deleteCompanyInterest(id)}
-            >
-              {({ openConfirmModal }) => (
-                <Icon onClick={openConfirmModal} name="trash" danger />
-              )}
-            </ConfirmModal>
-          </Flex>
-        ),
-      },
-    ];
-    const semesterOptions = [
-      {
-        value: 9999,
-        year: 9999,
-        semester: '',
-        label: 'Vis alle semestre',
-      },
-      ...this.props.semesters.map((semesterObj: CompanySemesterEntity) => {
-        const { id, year, semester } = semesterObj;
-        return {
-          id,
-          value: year,
-          year,
-          semester,
-          label: semesterToText({ ...semesterObj, language: 'norwegian' }),
-        };
-      }),
-    ].sort((o1, o2) => {
-      return Number(o1.year) === Number(o2.year)
-        ? o1.semester === 'spring'
-          ? -1
-          : 1
-        : Number(o1.year) > Number(o2.year)
+  const columns = [
+    {
+      title: 'Bedriftsnavn',
+      dataIndex: 'companyName',
+      render: (companyName: string, companyInterest: Record<string, any>) => (
+        <Link to={`/companyInterest/${companyInterest.id}/edit`}>
+          {companyInterest.company ? companyInterest.company.name : companyName}
+        </Link>
+      ),
+    },
+    {
+      title: 'Kontaktperson',
+      dataIndex: 'contactPerson',
+      render: (contactPerson: string) => <span>{contactPerson}</span>,
+    },
+    {
+      title: 'E-post',
+      dataIndex: 'mail',
+      render: (mail: string) => <span>{mail}</span>,
+    },
+    {
+      title: '',
+      dataIndex: 'id',
+      render: (id) => (
+        <Flex justifyContent="center">
+          <ConfirmModal
+            title="Bekreft fjerning av interesse"
+            message={'Er du sikker på at du vil fjerne?'}
+            closeOnConfirm
+            onConfirm={() => dispatch(deleteCompanyInterest(id))}
+          >
+            {({ openConfirmModal }) => (
+              <Icon onClick={openConfirmModal} name="trash" danger />
+            )}
+          </ConfirmModal>
+        </Flex>
+      ),
+    },
+  ];
+
+  const semesterOptions = [
+    {
+      value: 9999,
+      year: 9999,
+      semester: '',
+      label: 'Vis alle semestre',
+    },
+    ...semesters.map((semesterObj: CompanySemesterEntity) => {
+      const { id, year, semester } = semesterObj;
+      return {
+        id,
+        value: year,
+        year,
+        semester,
+        label: semesterToText({ ...semesterObj, language: 'norwegian' }),
+      };
+    }),
+  ].sort((o1, o2) => {
+    return Number(o1.year) === Number(o2.year)
+      ? o1.semester === 'spring'
         ? -1
-        : 1;
-    });
+        : 1
+      : Number(o1.year) > Number(o2.year)
+      ? -1
+      : 1;
+  });
 
-    return (
-      <Content>
-        <ListNavigation title="Bedriftsinteresser" />
-        <Flex
-          wrap
-          justifyContent="space-between"
-          alignItems="flex-end"
-          className={styles.section}
-        >
-          <Flex column>
-            <p>
-              Her finner du all praktisk informasjon knyttet til
-              <strong> bedriftsinteresser</strong>.
-            </p>
-            <Select
-              name="form-semester-selector"
-              value={this.props.selectedSemesterOption}
-              onChange={this.handleSemesterChange}
-              options={semesterOptions}
-              isClearable={false}
-              theme={selectTheme}
-              styles={selectStyles}
-            />
-          </Flex>
-          <Link to="/companyInterest/semesters">
-            <Button>Endre aktive semestre</Button>
-          </Link>
-          <Link to="/companyInterest/create">
-            <Button>Opprett ny bedriftsinteresse</Button>
-          </Link>
+  return (
+    <Content>
+      <ListNavigation title="Bedriftsinteresser" />
+      <Flex
+        wrap
+        justifyContent="space-between"
+        alignItems="flex-end"
+        className={styles.section}
+      >
+        <Flex column>
+          <p>
+            Her finner du all praktisk informasjon knyttet til
+            <strong> bedriftsinteresser</strong>.
+          </p>
+          <Select
+            name="form-semester-selector"
+            value={selectedSemesterOption}
+            onChange={handleSemesterChange}
+            options={semesterOptions}
+            isClearable={false}
+            theme={selectTheme}
+            styles={selectStyles}
+          />
+        </Flex>
+        <Link to="/companyInterest/semesters">
+          <Button>Endre aktive semestre</Button>
+        </Link>
+        <Link to="/companyInterest/create">
+          <Button>Opprett ny bedriftsinteresse</Button>
+        </Link>
+      </Flex>
+
+      <Flex
+        wrap
+        justifyContent="space-between"
+        alignItems="flex-end"
+        className={styles.section}
+      >
+        <Flex column>
+          <Select
+            name="form-event-selector"
+            value={selectedEventOption}
+            onChange={handleEventChange}
+            options={EVENT_TYPE_OPTIONS}
+            isClearable={false}
+            theme={selectTheme}
+            styles={selectStyles}
+          />
         </Flex>
 
-        <Flex
-          wrap
-          justifyContent="space-between"
-          alignItems="flex-end"
-          className={styles.section}
-        >
-          <Flex column>
-            <Select
-              name="form-event-selector"
-              value={this.props.selectedEventOption}
-              onChange={this.handleEventChange}
-              options={EVENT_TYPE_OPTIONS}
-              isClearable={false}
-              theme={selectTheme}
-              styles={selectStyles}
-            />
-          </Flex>
-
-          {generatedCSV ? (
-            <a href={generatedCSV.url} download={generatedCSV.filename}>
-              Last ned
-            </a>
-          ) : (
-            <Tooltip
-              style={
-                this.props.selectedSemesterOption.year
-                  ? { display: 'none' }
-                  : undefined
-              }
-              content={'Vennligst velg semester'}
+        {generatedCSV ? (
+          <a href={generatedCSV.url} download={generatedCSV.filename}>
+            Last ned
+          </a>
+        ) : (
+          <Tooltip
+            style={
+              selectedSemesterOption.year ? { display: 'none' } : undefined
+            }
+            content={'Vennligst velg semester'}
+          >
+            <Button
+              onClick={async () => setGeneratedCSV(await exportInterestList())}
+              disabled={!selectedSemesterOption.year}
             >
-              <Button
-                onClick={async () =>
-                  this.setState({
-                    generatedCSV: await this.exportInterestList(
-                      this.props.selectedEventOption.value
-                    ),
-                  })
-                }
-                disabled={!this.props.selectedSemesterOption.year}
-              >
-                Eksporter til CSV
-              </Button>
-            </Tooltip>
-          )}
-        </Flex>
+              Eksporter til CSV
+            </Button>
+          </Tooltip>
+        )}
+      </Flex>
 
-        <Table
-          columns={columns}
-          onLoad={(filters, sort) => {
-            this.props.fetch({
+      <Table
+        columns={columns}
+        onLoad={(filters) => {
+          dispatch(
+            fetch({
               next: true,
               filters,
-            });
-          }}
-          onChange={(filters, sort) => {
-            this.props.fetch({
+            })
+          );
+        }}
+        onChange={(filters) => {
+          dispatch(
+            fetch({
               filters,
-            });
-          }}
-          hasMore={this.props.hasMore}
-          loading={this.props.fetching}
-          data={this.props.companyInterestList}
-        />
-      </Content>
-    );
-  }
-}
+            })
+          );
+        }}
+        hasMore={hasMore}
+        loading={fetching}
+        data={companyInterestList}
+      />
+    </Content>
+  );
+};
 
 export default CompanyInterestList;
